Move complete node failure test to workflow display tests
- Move test from test_code_execution_node.py to test_workflow_display.py
- Use BaseNode instead of CodeExecutionNode for simpler test approach
- Fix mypy errors by importing JsonObject from correct module
- Test verifies edges are pruned when node completely fails to serialize
- Addresses feedback to use simpler BaseNode approach in workflow tests
